refactor: Refactor line drawing to function on two point click
Previously you would click 'Draw Line' in the menu options then
place two points on the maps and a line would be drawn between
them.

This was a little annoying when your point was already on the map,
now to draw a line you just select 'Draw Line' and click any two
locations, a new marker will not be placed but a line is drawn.